Fixed passing of username between processes.
In PowerShell/openssh-portable#327 there were some optimizations and the
way how Authctxt is created in forked process, but the change introduced
a bug - the user name sent between processes is `auth->pw->pw_name`, while
the received one is stored to `auth->user`. This fix changes it to send
`auth->user` as expected.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>