Update _PATH_TTY to point to the console
Currently openssh tests for a TTY by opening /dev/tty which does not exist on Windows. It most cases this then falls back to a prompt that works, but if the DISPLAY env var is set, then it tries to spawn "/usr/X11R6/bin/ssh-askpass" and fails - see issue #966, #1515, #762, #1088, #1356, and probably others.

Possible workarounds are to create a file `C:\dev\tty` so that it is able to open it, or to unset the DISPLAY variable, but it would be better if OpenSSH could detect this is a valid console. Code is from https://github.com/PowerShell/Win32-OpenSSH/issues/966#issuecomment-461989704