Try to resolve user SID also when it has same name as computer/domain.
This is a rework of the fix to allow using user name same as computer name.
Instead of trying to work with fully-qualified user names <domain>\<name>
(<name>\<name> in this case), fix this in the SID lookup function directly.
When the SID is a domain, simply try to resolve <name>\<name>.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>